+++ extensions/instagram/src/main/java/app/morphe/extension/instagram/utils/Pref.java
@@ -9,6 +9,7 @@ package app.morphe.extension.instagram.utils;
 
 import app.morphe.extension.instagram.settings.Settings;
 import app.morphe.extension.instagram.settings.SettingsStatus;
+import app.morphe.extension.crimera.SharedPref;
 
 @SuppressWarnings("unused")
 public class Pref {
@@ -43,24 +44,32 @@ public class Pref {
         return SharedPref.getBooleanPref(Settings.SANITIZE_SHARE_LINKS);
     }
 
+    public static boolean getTurnOnAllGhostModes() {
+        return SharedPref.getBooleanPref(Settings.TURN_ON_ALL_GHOST_MODES);
+    }
+
+    public static boolean setTurnOnAllGhostModes(boolean bool) {
+        return SharedPref.setBooleanPref(Settings.TURN_ON_ALL_GHOST_MODES.key,bool);
+    }
+
     public static boolean viewStoriesAnonymously() {
-        return SharedPref.getBooleanPref(Settings.VIEW_STORIES_ANONYMOUSLY);
+        return (SharedPref.getBooleanPref(Settings.VIEW_STORIES_ANONYMOUSLY) && SettingsStatus.viewStoriesAnonymously) || Pref.getTurnOnAllGhostModes();
     }
 
     public static boolean viewLiveAnonymously() {
-        return SharedPref.getBooleanPref(Settings.VIEW_LIVE_ANONYMOUSLY);
+        return (SharedPref.getBooleanPref(Settings.VIEW_LIVE_ANONYMOUSLY) && SettingsStatus.viewLiveAnonymously) || Pref.getTurnOnAllGhostModes();
     }
 
     public static boolean disableScreenshotDetection() {
-        return SharedPref.getBooleanPref(Settings.DISABLE_SCREENSHOT_DETECTION);
+        return SharedPref.getBooleanPref(Settings.DISABLE_SCREENSHOT_DETECTION) || Pref.getTurnOnAllGhostModes();
     }
 
     public static boolean disableTypingStatus() {
-        return SharedPref.getBooleanPref(Settings.DISABLE_TYPING_STATUS);
+        return SharedPref.getBooleanPref(Settings.DISABLE_TYPING_STATUS) || Pref.getTurnOnAllGhostModes();
     }
 
     public static boolean viewDmAnonymously() {
-        return SharedPref.getBooleanPref(Settings.VIEW_DM_ANONYMOUSLY);
+        return SharedPref.getBooleanPref(Settings.VIEW_DM_ANONYMOUSLY) || Pref.getTurnOnAllGhostModes();
     }
 
     public static boolean disableVideoAutoplay() {
@@ -175,6 +184,10 @@ public class Pref {
 
     public static boolean downloadUsernameFolder() {
         return SharedPref.getBooleanPref(Settings.DOWNLOAD_USERNAME_FOLDER);
+    }
+
+    public static String getCustomDownloadPath() {
+        return SharedPref.getStringPref(Settings.CUSTOM_DOWNLOAD_PATH);
     }
 
     public static boolean hideNavigationFeed() {