+++ extensions/instagram/stub/src/main/java/com/instagram/igds/components/button/IgdsButton.java
@@ -6,5 +6,5 @@ import android.widget.FrameLayout;
 public class IgdsButton extends FrameLayout{
     public IgdsButton(Context context){super(context);};
 
-    public void setText(String str){}
+    public void setText(CharSequence str){}
 }