+++ extensions/shared/library/src/main/java/app/morphe/extension/crimera/downloader/DownloadRequest.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright (C) 2026 piko <https://github.com/crimera/piko>
+ *
+ * This file is part of piko.
+ *
+ * Any modifications, derivatives, or substantial rewrites of this file
+ * must retain this copyright notice and the piko attribution
+ * in the source code and version control history.
+ */
+
+
+package app.morphe.extension.crimera.downloader;
+
+public class DownloadRequest {
+    public String url;
+    public String subFolder; 
+    public String fileName;
+
+    public DownloadRequest(String url, String subFolder, String fileName) {
+        this.url = url;
+        this.subFolder = subFolder;
+        this.fileName = fileName;
+    }
+}