+++ patches/src/main/kotlin/app/crimera/patches/twitter/misc/shareMenu/fingerprints/ActionEnumsFingerprint.kt
@@ -6,16 +6,21 @@
 
 package app.crimera.patches.twitter.misc.shareMenu.fingerprints
 
+import app.crimera.patches.twitter.utils.is_11_92_stable_or_greater
 import app.crimera.utils.InitMethod
 import app.crimera.utils.indexOfLastFilledNewArrayRange
 import app.crimera.utils.indexOfLastNewInstance
 import app.crimera.utils.instructionToString
 import app.morphe.patcher.Fingerprint
+import app.morphe.patcher.extensions.InstructionExtensions.addInstructions
+import app.morphe.patcher.extensions.InstructionExtensions.instructions
 import app.morphe.patcher.patch.BytecodePatchContext
 import app.morphe.patcher.util.proxy.mutableTypes.MutableField.Companion.toMutable
 import app.morphe.patcher.util.proxy.mutableTypes.MutableMethod
 import app.morphe.patcher.util.smali.toInstruction
 import app.morphe.patcher.util.smali.toInstructions
+import app.morphe.util.registersUsed
+import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.builder.MethodImplementationBuilder
 import com.android.tools.smali.dexlib2.iface.instruction.formats.Instruction21c
 import com.android.tools.smali.dexlib2.iface.instruction.formats.Instruction3rc
@@ -26,17 +31,21 @@ internal object ActionEnumsFingerprint : Fingerprint(
 )
 
 context(patchContext: BytecodePatchContext)
-fun addAction(name: String): String {
+fun addActionV1(name: String): String {
+    val classDef = ActionEnumsFingerprint.classDef
+    val classFields = classDef.fields
+    val classMethods = classDef.methods
+
     val field =
-        ActionEnumsFingerprint.classDef.fields
+        classFields
             .last()
             .toMutable()
     field.name = name
-    ActionEnumsFingerprint.classDef.fields.add(field)
+    classFields.add(field)
 
-    // TODO: handle nulls
-    val oldInit = ActionEnumsFingerprint.classDef.methods.first { it.name == "<clinit>" }
+    val methodName = "<clinit>"
 
+    val oldInit = classMethods.first { it.name == methodName }
     val oldInitInstructions = oldInit.implementation!!.instructions.toMutableList()
 
     val lastNewInstance = (oldInitInstructions[oldInitInstructions.indexOfLastNewInstance] as Instruction21c)
@@ -44,9 +53,9 @@ fun addAction(name: String): String {
     val stringReg = lastNewInstanceReg + 1
     val constReg = stringReg + 1
 
-    val fieldCall = "${lastNewInstance.reference}->$name:${lastNewInstance.reference}"
-
     val newArrayPos = oldInitInstructions.indexOfLastFilledNewArrayRange
+
+    val fieldCall = "${lastNewInstance.reference}->$name:${lastNewInstance.reference}"
 
     val arr = oldInitInstructions[newArrayPos] as Instruction3rc
     val arrStartReg = arr.startRegister
@@ -100,8 +109,103 @@ fun addAction(name: String): String {
             implementation = newInitImplementation.methodImplementation,
         )
 
-    val actionsEnumClass = ActionEnumsFingerprint.classDef
-    actionsEnumClass.methods.removeIf { it.name == "<clinit>" }
-    actionsEnumClass.methods.add(MutableMethod(newInit))
+    classMethods.removeIf { it.name == methodName }
+    classMethods.add(MutableMethod(newInit))
     return fieldCall
+}
+
+context(patchContext: BytecodePatchContext)
+fun addActionV2(name: String): String {
+    val classDef = ActionEnumsFingerprint.classDef
+    val classFields = classDef.fields
+    val classMethods = classDef.methods
+
+    val fieldCall = "$classDef->$name:$classDef"
+
+    val field =
+        classFields
+            .last()
+            .toMutable()
+    field.name = name
+    classFields.add(field)
+
+    val methodName = "\$values"
+    val oldValues = classMethods.first { it.name == methodName }
+
+    val oldValuesInstructions = oldValues.implementation!!.instructions.toMutableList()
+
+    val newArrayPos = oldValuesInstructions.indexOfLastFilledNewArrayRange
+
+    val arr = oldValuesInstructions[newArrayPos] as Instruction3rc
+    val arrStartReg = arr.startRegister
+    val arrEndReg = (arrStartReg + arr.registerCount)
+
+    ActionEnumsFingerprint.method.apply {
+        val lastNewInstance = instructions.last { it.opcode == Opcode.NEW_INSTANCE }
+        val lastNewInstanceReg = lastNewInstance.registersUsed[0]
+        val stringReg = lastNewInstanceReg + 1
+        val constReg = stringReg + 1
+
+        val lastInvokeStaticIndex = instructions.last { it.opcode == Opcode.INVOKE_STATIC }.location.index
+        addInstructions(
+            lastInvokeStaticIndex,
+            """
+            new-instance v$lastNewInstanceReg, $classDef
+            const-string v$stringReg, "$name"
+            const/16 v$constReg, ${arrEndReg - arrStartReg}
+            invoke-direct {v$lastNewInstanceReg, v$stringReg, v$constReg}, $classDef-><init>(Ljava/lang/String;I)V
+            sput-object v$lastNewInstanceReg, $fieldCall
+            """.trimIndent(),
+        )
+    }
+
+    oldValuesInstructions[newArrayPos] =
+        """
+        filled-new-array/range {v$arrStartReg .. v${arrEndReg + arrStartReg}}, [$classDef
+        """.trimIndent().toInstruction()
+
+    oldValuesInstructions.add(
+        newArrayPos,
+        """
+        sget-object v$arrEndReg, $fieldCall
+        """.trimIndent().toInstruction(),
+    )
+
+    val instructions =
+        oldValuesInstructions.map { ins ->
+            instructionToString(ins)
+        }
+
+    val newInitImplementation = MethodImplementationBuilder(arrEndReg + 1)
+    instructions.forEach { ins ->
+        newInitImplementation.addInstruction(ins.toInstruction())
+    }
+
+    val newInit =
+        InitMethod(
+            validator = { oldValues.validateReference() },
+            compare = { oldValues.compareTo(it) },
+            definingClass = oldValues.definingClass,
+            name = oldValues.name,
+            parameterTypes = oldValues.parameterTypes,
+            returnType = oldValues.returnType,
+            annotations = oldValues.annotations,
+            accessFlags = oldValues.accessFlags,
+            hiddenApiRestrictions = oldValues.hiddenApiRestrictions,
+            parameters = oldValues.parameters,
+            implementation = newInitImplementation.methodImplementation,
+        )
+
+    classMethods.removeIf { it.name == methodName }
+    classMethods.add(MutableMethod(newInit))
+    return fieldCall
+}
+
+context(patchContext: BytecodePatchContext)
+fun addAction(name: String): String {
+    if (is_11_92_stable_or_greater) {
+        return addActionV2(name)
+    } else {
+        return addActionV1(name)
+    }
 }